+++ newsbycountry/apps/references/models.py
@@ -15,3 +15,4 @@ class Tags(models.Model):
 
     def __str__(self):
         return self.name
+
